Massive rewrite.
New logic is similar to id3-frame.lisp in that it uses find-class to instantiate atoms.

There are some potential problems with this: first, the ILST DATA atoms are a potential
problem area in that there are generically named.  For now, I'm not reading any container
atoms that have DATA atoms in them (other than ILST), but ...

The second potential problem is that I'm ignoring case when doing find-symbol and
some atoms (eg "aArt") are case-sensitive.